requirements.txt
discord.py>=2.3.0
yfinance>=0.2.30
pytz>=2023.3
aiohttp>=3.9.0
python-dotenv>=1.0.0
